fix: request individual cellxgene collections to avoid timeout on datasets (#2465) (#2466)
