+++ next/utils/atlasPages.ts
@@ -12,7 +12,11 @@ import { ProjectsResponse } from "../apis/azul/hca-dcp/common/responses";
 import { processEntityValue } from "../apis/azul/hca-dcp/common/utils";
 import { config } from "../config/config";
 import { NETWORKS } from "../constants/networks";
-import { processAtlas, processNetwork } from "./network";
+import {
+  fetchCXGDatasetsForAtlases,
+  processAtlas,
+  processNetwork,
+} from "./network";
 
 interface StaticPaths extends ParsedUrlQuery {
   atlas: string;
@@ -68,17 +72,8 @@ export async function getContentStaticProps(
     }
   }
 
-  const cxgDatasets = [];
-  if (atlas.cxgId) {
-    const response = await fetch(
-      `https://api.cellxgene.cziscience.com/curation/v1/collections/${atlas.cxgId}`
-    );
-    const cxgCollection = await response.json();
-    cxgDatasets.push(
-      ...mapDatasets(cxgCollection.datasets, cxgCollection.collection_id)
-    );
-    cxgDatasets.sort(sortCXGDatasets);
-  }
+  const cxgDatasets = await fetchCXGDatasetsForAtlases([atlas]);
+  cxgDatasets.sort(sortCXGDatasets);
 
   return {
     props: {
@@ -88,19 +83,6 @@ export async function getContentStaticProps(
       projectsResponses,
     },
   };
-}
-
-/**
- * Returns CELLxGENE datasets with the corresponding collection ID.
- * @param cxgDatasets - CELLxGENE dataset responses.
- * @param collection_id - Collection ID.
- * @returns CELLxGENE datasets.
- */
-function mapDatasets(
-  cxgDatasets: Omit<CXGDataset, "collection_id">[],
-  collection_id: string
-): CXGDataset[] {
-  return cxgDatasets.map((cxgDataset) => ({ ...cxgDataset, collection_id }));
 }
 
 /**
